+++ tests/helper/utils.py
@@ -2,8 +2,6 @@ import os
 import re
 
 def get_package_list(client):
-    """Return list with the installed packages.
-    Needs the fixture client to connect into the image"""
     (exit_code, output, error) = client.execute_command("dpkg -l")
     assert exit_code == 0, f"no {error=} expected"
     pkgslist = []
@@ -16,43 +14,16 @@ def get_package_list(client):
     return pkgslist
 
 
-def read_test_config(features, testname, suffix = ".list"):
-    """Collect the configuration of a test from all enabled features.
-    Needs the list of enabled features and the name of the test, the suffix is
-    optional. It returns a list of the aggregated configs for a test."""
+def read_test_config(features, testname, suffix):
     config = []
     for feature in features:
-        path = (f"/gardenlinux/features/{feature}/test/{testname}.d/" +
-            f"{testname}{suffix}")
+        path = ("/gardenlinux/features/%s/test/%s.d/%s%s" % (feature, testname, testname, suffix))
         if os.path.isfile(path):
             file = open(path, 'r')
             for line in file:
-                # Skip comment lines
-                if re.match(r'^ *#',line):
+                if line.startswith('#'):
                     continue
-                # Skip empty lines
                 if re.match(r'^\s*$', line):
                     continue
                 config.append(line.strip('\n'))
     return config
-
-
-def disabled_by(features, testname):
-    """Checks is a test is explicitly disabled by a feature.
-    Needs the list of enabled features and the name of the test.
-    It returns a list of the features where the test is disabled."""
-    disabled = []
-    for feature in features:
-        path = (f"/gardenlinux/features/{feature}/test/{testname}.disable")
-        if os.path.isfile(path):
-            disabled.append(feature)
-    return disabled
-
-
-class AptUpdate():
-    def __new__(cls, client):
-        if not hasattr(cls, 'instance'):
-            cls.instance = super(AptUpdate, cls).__new__(cls)
-
-        (exit_code, output, error) = client.execute_command("apt-get update")
-        assert exit_code == 0, f"no {error=} expected"
